Board mode and Cell signal connection
The Board now has a mode so we can process things properly based on the board mode / game state, its initial selections are in based on what I think we'll use to start.
The Board now connects to the cell pressed signal, so it knows when a specific cell has been clicked and can respond accordingly in the future.