First interaction and reset of board
You can now click on the board to change a cell to the Base state (like placing your base that will be targeted later). Then you can press the reset button to reset the board to clear it. Just to show the change happening, the board is initialized as 5x5 but when you press reset it changes it to 10x10, where it will stay for now every time you reset it.